requirements.txt
numpy==2.2.3
opencv_python==4.10.0.84
Pillow==11.1.0
PySide6==6.8.1.1
PySide6==6.8.2.1
PySide6_Addons==6.8.1.1
PySide6_Essentials==6.8.1.1
PyYAML==6.0.2
PyYAML==6.0.2
torch==2.5.1+cu124
ultralytics==8.3.58
